Move helper function to libultraship to (hopefully) get Jenkins happy
