+++ soh/soh/Enhancements/debugger/ImGuiHelpers.cpp
@@ -1,38 +1,11 @@
 #include "ImGuiHelpers.h"
-
-// Automatically add newlines to break up tooltips longer than a specified number of characters
-// Manually included newlines will still be respected and reset the line length
-std::string BreakTooltip(const char* text) {
-    const int lineLength = 60;
-    std::string newText = std::string(text);
-    const int tipLength = newText.length();
-    int lastSpace = -1;
-    int currentLineLength = 0;
-    for (int currentCharacter = 0; currentCharacter < tipLength; currentCharacter++) {
-        if (newText[currentCharacter] == '\n') {
-            currentLineLength = 0;
-            lastSpace = -1;
-            continue;
-        } 
-        else if (newText[currentCharacter] == ' ') {
-            lastSpace = currentCharacter;
-        }
-        
-        if ((currentLineLength >= lineLength) && (lastSpace >= 0)) {
-            newText[lastSpace] = '\n';
-            currentLineLength = currentCharacter - lastSpace - 1;
-            lastSpace = -1;
-        }
-        currentLineLength++;
-    }
-    return newText;
-}
+#include "../../../../libultraship/libultraship/ImGuiImpl.h"
 
 // Adds a text tooltip for the previous ImGui item
 void SetLastItemHoverText(const std::string& text) {
     if (ImGui::IsItemHovered()) {
         ImGui::BeginTooltip();
-        ImGui::Text(BreakTooltip(text.c_str()).c_str());
+        ImGui::Text(SohImGui::BreakTooltip(text.c_str()).c_str());
         ImGui::EndTooltip();
     }
 }
@@ -43,7 +16,7 @@ void InsertHelpHoverText(const std::string& text) {
     ImGui::TextColored(ImVec4(0.7f, 0.7f, 0.7f, 1.0f), "?");
     if (ImGui::IsItemHovered()) {
         ImGui::BeginTooltip();
-        ImGui::Text(BreakTooltip(text.c_str()).c_str());
+        ImGui::Text(SohImGui::BreakTooltip(text.c_str()).c_str());
         ImGui::EndTooltip();
     }
 }
