+++ soh/soh/Enhancements/debugger/ImGuiHelpers.h
@@ -3,8 +3,6 @@
 
 #include <string>
 
-std::string BreakTooltip(const char* text);
-
 void SetLastItemHoverText(const std::string& text);
 
 void InsertHelpHoverText(const std::string& text);
